+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_up.py
@@ -23,8 +23,17 @@ class ContainerAppUpImageTest(ScenarioTest):
 
     @live_only()
     @ResourceGroupPreparer(location="eastus2")
-    def test_containerapp_up_source_with_buildpack_e2e(self, resource_group):
-        source_path = os.path.join(TEST_DIR, os.path.join("data", "source_built_using_buildpack"))
+    def test_containerapp_up_source_with_bullseye_buildpack_e2e(self, resource_group):
+        source_path = os.path.join(TEST_DIR, os.path.join("data", "source_built_using_bullseye_buildpack_net7"))
+        ingress = 'external'
+        target_port = '8080'
+        create_and_verify_containerapp_up(self,resource_group=resource_group, source_path=source_path, ingress=ingress, target_port=target_port)
+
+
+    @live_only()
+    @ResourceGroupPreparer(location="eastus2")
+    def test_containerapp_up_source_with_bookworm_buildpack_e2e(self, resource_group):
+        source_path = os.path.join(TEST_DIR, os.path.join("data", "source_built_using_bookworm_buildpack_net8"))
         ingress = 'external'
         target_port = '8080'
         create_and_verify_containerapp_up(self,resource_group=resource_group, source_path=source_path, ingress=ingress, target_port=target_port)