+++ src/scvmm/azext_scvmm/vendored_sdks/scvmm/aio/operations/_operations.py
@@ -29,7 +29,7 @@ class Operations:
     instantiates it for you and attaches it as an attribute.
 
     :ivar models: Alias to model classes used in this operation group.
-    :type models: ~scvmm.models
+    :type models: ~azure.mgmt.scvmm.models
     :param client: Client for service requests.
     :param config: Configuration of service client.
     :param serializer: An object model serializer.
@@ -48,18 +48,17 @@ class Operations:
     def list(
         self,
         **kwargs: Any
-    ) -> AsyncIterable["_models.ResourceProviderOperationList"]:
+    ) -> AsyncIterable["_models.OperationListResult"]:
         """Returns list of all operations.
 
         :keyword callable cls: A custom type or function that will be passed the direct response
-        :return: An iterator like instance of either ResourceProviderOperationList or the result of
-         cls(response)
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[~scvmm.models.ResourceProviderOperationList]
+        :return: An iterator like instance of either OperationListResult or the result of cls(response)
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.mgmt.scvmm.models.OperationListResult]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
-        api_version = kwargs.pop('api_version', "2023-04-01-preview")  # type: str
+        api_version = kwargs.pop('api_version', "2023-10-07")  # type: str
 
-        cls = kwargs.pop('cls', None)  # type: ClsType["_models.ResourceProviderOperationList"]
+        cls = kwargs.pop('cls', None)  # type: ClsType["_models.OperationListResult"]
         error_map = {
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
@@ -86,7 +85,7 @@ class Operations:
             return request
 
         async def extract_data(pipeline_response):
-            deserialized = self._deserialize("ResourceProviderOperationList", pipeline_response)
+            deserialized = self._deserialize("OperationListResult", pipeline_response)
             list_of_elem = deserialized.value
             if cls:
                 list_of_elem = cls(list_of_elem)